added logger
